requirements.txt
Flask
requests
gunicorn
python-pptx
...
